refactor: blackJackRule 제거
