--- src/main/java/blackjack/domain/blackjack/BlackJackRule.java
@@ -1,9 +1,0 @@
-package blackjack.domain.blackjack;
-
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Player;
-
-public interface BlackJackRule {
-
-    ResultType calculateDealerResult(Dealer dealer, Player player);
-}
